Add the ability to turn off some diagnostic and optimization options
Some of our error reporting, assertions, and diagnostics can interfere
with the ability of the compiler to make certain optimizations. So
now you have the option to turn that off.

* `DEBUG=1` now only turns on debugging symbol options that for sure won't cause any execution slowdown.
* `DEBUG_EXTRA=1` will turn on some even more extra debugging information but may cause execution slowdowns in the form of missed optimization opportunities.
* `FAST_CFLAGS_BUT_NO_VERIFICATION=1` will turn off any sort of runtime assertions at all, which should allow even more speed but you get little protection against
  undefined behavior if anything goes wrong, and even less information about what happened if it does.

Also disables by default one particular optimization that may (or may not)
be interfering with other optimizations, so I am disabling it by default
as well.

`EXPERIMENTAL_OPTIMIZATIONS=1` can turn them back on