Bump the version to `1.14-TechSY730` from upstream's `1.13`
The `-TechSY730` just indicates a prerelease specifier, it doesn't affect version comparison,
and it doesn't change what server you upload to.
This is to distinguish it from a hypothetical `1.14` coming from upstream in the future
(seems unlikely as the project is dormant).
You can remove this is you don't feel comfortable advertising my fork.