Instead of emptying out the legalMoves queue from the front, go from the back.
Doing it from the front prompted moving everything down every remove,
when all (or all but one) is going away anyways. Now it goes from the
back to the forward, saving all of those shifts.

When just deleting one legal move, then is will continue to shift
to fill the gap.